refactor: move some schemas & types and change file names to be in sync
