+++ lib/src/commands/moderation/mute.ts
@@ -4,10 +4,10 @@ import Embeds from "../../util/constants/Embeds";
 import timestampToDuration from "../../util/functions/timestampToDuration";
 import Server from "../../mongo/model/Server";
 import canExecute from "../../util/functions/canExecute";
-import {LogType} from "../../mongo/schema/Log";
-import {IPunishment, toEmbedField} from "../../mongo/schema/Punishment";
+import {IPunishment, toEmbedField} from "../../mongo/schema/PunishmentSchema";
 import AuxdibotCommandInteraction from "../../util/templates/AuxdibotCommandInteraction";
 import GuildAuxdibotCommandData from "../../util/types/commandData/GuildAuxdibotCommandData";
+import {LogType} from "../../util/types/Log";
 
 const muteCommand = <AuxdibotCommand>{
     data: new SlashCommandBuilder()
