+++ lib/src/commands/permissions/permissions.ts
@@ -1,10 +1,10 @@
 import {SlashCommandBuilder} from "discord.js";
 import AuxdibotCommand from "../../util/templates/AuxdibotCommand";
 import Embeds from '../../util/constants/Embeds';
-import {IPermissionOverride} from "../../mongo/schema/PermissionOverride";
-import {LogType} from "../../mongo/schema/Log";
+import {IPermissionOverride} from "../../mongo/schema/PermissionOverrideSchema";
 import AuxdibotCommandInteraction from "../../util/templates/AuxdibotCommandInteraction";
 import GuildAuxdibotCommandData from "../../util/types/commandData/GuildAuxdibotCommandData";
+import {LogType} from "../../util/types/Log";
 
 const permissionsCommand = <AuxdibotCommand>{
     data: new SlashCommandBuilder()
