+++ lib/src/commands/settings/join_roles.ts
@@ -1,9 +1,9 @@
 import {SlashCommandBuilder, APIEmbed, Role, PermissionsBitField} from "discord.js";
 import AuxdibotCommand from "../../util/templates/AuxdibotCommand";
 import Embeds from "../../util/constants/Embeds";
-import {LogType} from "../../mongo/schema/Log";
 import AuxdibotCommandInteraction from "../../util/templates/AuxdibotCommandInteraction";
 import GuildAuxdibotCommandData from "../../util/types/commandData/GuildAuxdibotCommandData";
+import {LogType} from "../../util/types/Log";
 
 const joinRolesCommand = <AuxdibotCommand>{
     data: new SlashCommandBuilder()
