+++ lib/src/commands/settings/settings.ts
@@ -5,11 +5,9 @@ import {
 } from "discord.js";
 import AuxdibotCommand from "../../util/templates/AuxdibotCommand";
 import Embeds from "../../util/constants/Embeds";
-import {
-    LogType
-} from "../../mongo/schema/Log";
 import AuxdibotCommandInteraction from "../../util/templates/AuxdibotCommandInteraction";
 import GuildAuxdibotCommandData from "../../util/types/commandData/GuildAuxdibotCommandData";
+import {LogType} from "../../util/types/Log";
 
 const settingsCommand = < AuxdibotCommand > {
     data: new SlashCommandBuilder()
