+++ lib/src/mongo/schema/PunishmentSchema.ts
