+++ lib/src/mongo/schema/ReactionRoleSchema.ts
@@ -3,7 +3,7 @@ export type IReaction = { emoji: string, role: string };
 let ReactionSchema = new mongoose.Schema<IReaction>({
     emoji: { type: String, required: true },
     role: { type: String, required: true }
-})
+}, { _id: false })
 export interface IReactionRole {
     message_id: string;
     reactions: IReaction[];
@@ -12,5 +12,5 @@ export interface IReactionRole {
 const ReactionRoleSchema = new mongoose.Schema<IReactionRole>({
     message_id: { type: String, required: true },
     reactions: { type: [ReactionSchema], required: true },
-})
+}, { _id: false })
 export default ReactionRoleSchema;
